requirements.txt
requests~=2.28.1
fake-useragent
lxml~=4.9.1
pymongo~=4.3.2
SQLAlchemy~=1.4.42
mysqlclient
schedule~=1.1.0
PyYAML~=6.0